Add one semantic future and three bug futures about generic classes
The semantic is about how doing a query on a single field in a formal argument
makes all param fields generic even when they have default initializers.

The bugs are: A param real in a class/record makes the type name include a dot.
A record with only param and type fields doesn't get an emptyRecordPlaceholder
field added, so if it needs to be heap allocated, it fails with a 0 byte
allocation. A class with multiple param fields gets instantiated twice,
causing its type name to be "C(1,1)(3)" in an error message.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15221 3a8e244f-b0f2-452b-bcba-4c88e055c3ca