Corrected a few inaccuracies in how runtime details are reported
Most importantly, the number of blocked tasks no longer includes threads that
are idle.  Also, there is additional code to keep track of how many threads are
executing more than one task, e.g., when a thread that is at the end of a
cobegin begins executing one of the tasks inside the cobegin.  Finally, the
number of idle threads no longer includes those for which there is work to do,
but that haven't yet been so notified.  Additional locks are taken to ensure
the information reported is accurate at a particular point in time during
execution, rather than piecing information together from different points in
time.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15224 3a8e244f-b0f2-452b-bcba-4c88e055c3ca