added file name/line number info to verbose comm output
Made more primitives require file name and line number information so
that we can pass file names and line numbers to the comm_get and
comm_put functions.  Since this requires passing line numbers and file
names to the most basic of all primitives, PRIM_MOVE, I plan to look
into simplifying the code so that we always pass line number and file
name info, maybe so that we can keep track of the stack as well.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15227 3a8e244f-b0f2-452b-bcba-4c88e055c3ca