improved privatization of Block1D array class (14 to 8 remote gets!)
This checkin fixes the the domain over which the locArr array field of
the Block1DArr is declared.  It used to be declared over the
targetLocDom domain in the distribution class for the domain on locale
0.  It is now declared over the same domain in the distribution class
for the domain on the private locale.

Also simplified the privatization interface to require a single
function (privatize) rather than two (privatize1 and privatize2).

Updated the expected number of remote gets for streamCommCheck from 14
to 8, but increased the range by 1 to avoid the non-deterministic
regressions.  I still have to look into what is causing the
non-determinism.  I now only know that the main non-determinism (8 or
9 remote gets) is stemming from the remote thread invocation to signal
that the thread is complete.  I'm still unsure why there are 10 remote
gets because I haven't seen that outside of the test suite runs.  I
may have to use an optimized runtime to see this.

The addition of line numbers to the verbose comm output helped isolate
this issue.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15229 3a8e244f-b0f2-452b-bcba-4c88e055c3ca