renamed some module classes, methods and functions for consistency
This change unifies naming of array, domain, and distribution classes
so that they all end in the same way, with Arr, Dom, or Dist,
respectively.  Previously some were abbreviated and others were not.

Removed declaration of global variable Block from the standard
modules.  This variable was declared to be an instance of
DefaultArithmeticDist before distributions were really supported.
Removed uses of this variable from many test cases that embedded no-op
expressions like 'distributed Block' in the code.

Fixed an issue with my last checkin in which some test cases referred
to Distribution (now BaseDist) directly.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15235 3a8e244f-b0f2-452b-bcba-4c88e055c3ca