improved applicability of remote value forwarding
Relaxed the restriction on remote value forwarding so that it now
applies to classes, but not records or unions.  It used to be disabled
for all three kinds of types.

Added two more cases to codegen for chpl_add_to_task_list to handle
_EndCount in struct if it is not a reference.

Refactored the _upEndCount and _downEndCount functions in the modules
so that they use a nested function with the "dont disable remote value
forwarding" pragma so that remote value forwarding can apply to the
on-statements within these functions.

Updated the expected number of remote gets for streamCommCheck from 8
to 4, and removed the buffer that used to allow between 7 and 11.  It
now allows only for exactly 4.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15236 3a8e244f-b0f2-452b-bcba-4c88e055c3ca