Added "private" pragma to defaultDist object.
The defaultDist object is a global instance of the DefaultDist class
that is used to create undistributed or default arrays and domains.
Made this object "private" so that arrays and domains created on
locales other than Locale 0 will not have to make a remote reference
to this object.  On all locales other than Locale 0, this object is
constructed in ChapelLocale when each locale is constructed.

This decreases the number of remote gets for streamCommCheck if
tasksPerLocale is greater than one.  The number of remote gets is now
four times the number of tasks per locale times one less than the
number of locales (4 x TPL x NL).  These remaining four remote gets
are related to accessing the iterator class when building the follower
iterator class on different locales.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15237 3a8e244f-b0f2-452b-bcba-4c88e055c3ca