introduction of iterator records
This check-in makes a major change to the iterator transform,
introducing an iterator record.  Every iterator is still transformed
into an iterator class with methods, and now also an iterator record.
The _getIterator function is now custom to every iterator; it takes an
iterator record and returns an iterator class.

The iterator record captures the arguments and the iterator classes
are only created on invocation.  Before they were always copied on
invocation.  This should help to eliminate the last 3 remote gets in
the stream benchmark and make iterator classes much easier to
destroy/free.  It does not yet help with the remote gets because the
leader and follower iterator records are derived from the serial
iterator class rather than the serial iterator record.  I plan to make
that change soon.

I previously explored changing iterator classes into iterator records.
This doesn't work because recursive iterators require dynamic storage
allocation.  Also, our handling of references may not be robust enough
in the iterator transform.

I believe this check-in leaves some code in the compiler that is now
dead.  I plan to eliminate it soon.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15243 3a8e244f-b0f2-452b-bcba-4c88e055c3ca