simplified dynamic dispatch on iterators
With the introduction of iterator records, iterator dynamic dispatch
changed slightly.  This eliminates the old code to get the dynamic
iterator class (unnecessary now that the dynamic iterator class is
selected in the getIterator function).  Because that code chose now an
iterator record, it introduced a warning in the generated C code about
type-punned pointers.  This warning only showed up if CHPL_COMM_DEBUG
was set to false.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15245 3a8e244f-b0f2-452b-bcba-4c88e055c3ca