Make _chpl_comm_fork able to handle big arguments in GASNet
If the argument for a remote fork is bigger than allowed in a medium active
message, handle it by forking the remote task and doing a remote get at the
beginning of the new task.

Long AMs are only guaranteed by GASNet to handle at least 512 bytes, the same
size guaranteed for medium AMs. Also, long AMs require a buffer to be
allocated on the remote node to handle the argument.

Also, added (start|stop)VerboseCommHere and (start|stop)CommDiagnosticsHere.
These functions turn this functionality on or off locally, but do not
broadcast the change. This is useful when trying to analyze a loop that is
being executed by multiple locales, but not wanting to look at loop overhead.
For example:

forall x in blockDistArr {
  startCommDiagnosticsHere();
  ...
  stopCommDiagnosticsHere();
}

Made stubs for these functions in the three non-GASNet comm layers.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15246 3a8e244f-b0f2-452b-bcba-4c88e055c3ca