Applying a GASNet-supplied patch to the vapi/ibv conduit to better support xlc
Trying to compile GASNet for infiniband on an IBM using their xlc
compiler resulted in a warning/error in the use of non-constant
expressions within the offsetof() macro.  Reporting this problem
to the GASNet team resulted in the following patch which they
will include in their next release.  I documented this change in
our GASNet README file which tracks differences between our
GASNet code and the official 1.12.0 release.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15249 3a8e244f-b0f2-452b-bcba-4c88e055c3ca