Free memory allocated for iterator class copies
At the end of loops, the compiler now inserts a call to _freeIterator(), which
is a new function in ChapelArray, to free memory allocations for iterator class
copies.  This in turn simply calls chpl_free() on its argument, but only if it
is not a tuple.  (Tuples are not created by explicitly allocating memory.)
Special care was needed to ensure label statements were handled correctly in
build.cpp (due to adding the call to _freeIterator() at the end of loops after
the break label that is normally there), and to ensure scalar replacement was
not disabled in scalarReplace.cpp (due to having the iterator class copy as an
argument to chpl_free()).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15251 3a8e244f-b0f2-452b-bcba-4c88e055c3ca