Improving runtime's GASNet make rules
In the version of the Makefiles that have been checked in for quite
some time, the GASNet directory is rebuilt (= reconfigured + recompiled)
whenever the GASNet .mak file is missing.  This means that even doing
a "make clean" or "clobber" from an empty directory will build GASNet,
which seems unfortunate given how long it takes.  Improved these rules
so that if a clean or clobber rule is performed from the runtime
directory, GASNet will not be rebuilt.  I suspect that this isn't the
most elegant solution, but it was the only reasonable one I could come
up with in short order.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15253 3a8e244f-b0f2-452b-bcba-4c88e055c3ca