Removing hard-coding of spawner as ssh within infiniband gasnet launcher
For some reason while I was working on infiniband for the CX1, I hardcoded the
spawner to use ssh within the launch code.  This should really be set at
configuration time or via an environment variable, not hardcoded within the
sources.  This caused problems, for example, on the IBM Power6 cluster I was
using where MPI wants to be used as the launcher.  I need to get back to the
CX1 soon, so will deal with this on that platform then if it causes problems.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15254 3a8e244f-b0f2-452b-bcba-4c88e055c3ca