Added a new script to debug the CHPL_ environment
I added a new script -- printchplenv.pl -- that prints relevant
Chapel environment variables and their values, whether manually
set or inferred.  This should be helpful in debugging environments,
particularly as the number of environment variables increases.
Note that some environment variables are only printed out if they
are meaningful.  For example, CHPL_COMM_SUBSTRATE is only printed
if CHPL_COMM is not 'none' and CHPL_GASNET_SEGMENT is only printed
if CHPL_COMM is 'gasnet'.

Fixed a bug in the chplmake script that this new script revealed in
which setting CHPL_MAKE would not override the value that was used.

Updated README.chplenv to mention this script and README.bugs to
request that users send the output of this script along with any bug
reports.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15259 3a8e244f-b0f2-452b-bcba-4c88e055c3ca