Added code to free memory allocated when converting numbers to strings
A vector for constructor calls was added in insertDestructors(), which is
populated with constructor calls as well as calls to the primitive PRIM_CAST
if the argument is a number and the result is a string.  The return value of
this primitive is tracked just as if it were a value returned by a constructor.
In order for all this to work, the _real*_to_string() functions were modified
to always allocate a string, even when the value being converted is a NaN or
infinity, and chpl_free's argument is always cast to void* to avoid warnings
about strings losing their constness.  (Strings are const char*.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15261 3a8e244f-b0f2-452b-bcba-4c88e055c3ca