Made Lexical Structure changes in the implementation.
Disallowed identifiers that start with $.

Allowed binary and hex literals to be written with an uppercase B or X
as well as the lowercase.  The exponent e or E was already supported
in either case in the implementation.

Added compilerError and compilerWarning functions to the ChapelBase
module, removing their status as keywords.  This changes their
semantics slightly.  Whereas we used to allow types to be passed to
them, we now require parameters only.  This is because there is no way
in Chapel to specify a variable number of arguments where some of the
arguments are types and some are parameters.

To support printing types in compilerError and compilerWarning
function calls, I added a new function typeToString that takes a type
and returns a string parameter.  Documentation forthcoming.  Added
PRIM_TYPE_TO_STRING to support this function.

Updated compilerError and compilerWarning calls so that their actuals
are always parameter strings.  Other parameters are not yet supported.

Made changes to build.cpp since PRIM_ERROR and PRIM_WARNING must now
be used only in the functions in the ChapelBase module.  I changed
instances of this primitive to instances of the function call
"compilerError" instead.  This makes handling of these errors and
warnings easier because the function arguments can be looked at rather
than the actuals which go through a tuple because of the variable
argument list.

Updated tests that used $ at the beginning of identifiers and compiler
warnings/errors with type arguments.

I tried to allow support of real literals like 3. but ran into even
more serious issues because of the .. operator.  If we support such
literals, it seems that the range 3..7 is interpreted instead as the
real literal 3. followed by a member access . followed by an
unexpected 7.  I tried to work with right context in lex but couldn't
make this work and I think that may be more complicated than we want.
So my current plan is to change the spec note to mention the range
operator of .. as the reason why we require a digit to follow the
decimal point as well as the e method business.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15269 3a8e244f-b0f2-452b-bcba-4c88e055c3ca