an interesting .locale on distributed arrays of strings bug
This test shows an interesting bug in which .locale applied to the
elements of a distributed array of strings returns the locale on
which the RHS string values were stored (or maybe the locale on
which the assignment occurred?) rather than the locale on which
the array element is stored.  This means that querying the .locale
of all the array elements is correct when the array is uninitialized,
but incorrect if locale 0 assigns all of the array elements to a
string literal stored on locale 0.

(Note that this test only fails on multilocale runs like GASNet).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15285 3a8e244f-b0f2-452b-bcba-4c88e055c3ca