Fixed the leader/follower iterators for my distributed associative domain/array
I had written approximately correct implementations of leader/follower
iterators for the associative domain and array, but not correct
enough to support the correct execution of a forall loop.

It's worth noting that my current leader/follower iterators aren't
really appropriate for zippered iteration unless all the expressions
share the same domain.  I'm not actually certain what zippered
iteration ought to mean for associative domains/arrays (not to
mention sparse, opaque, etc.)  It may be that we should define the
semantics in terms of index values rather than simply shape since
"shape" is not as well-defined in this range.

[Steve, note that the above is one of the reasons I've been hoping
to get a "standalone parallel iterator" implemented -- it'd be
nice to be able to implement such an iterator for the associative
domain/array and make the leader/follower stuff result in a halt
for now until we are able to come up with a natural and implementable
semantic for it.  The current use of leader/follower is very
artificial in this case given that it will likely break if used
for anything other than the standalone case.]




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15286 3a8e244f-b0f2-452b-bcba-4c88e055c3ca