Add a function to compute integer exponentiation.
For integer exponentiation, instead of calling the C pow function, which takes
two doubles and returns a double, implement it as a function in ChapelBase.

Update parameter folding of integer exponentiation so 0 ** negative_number is
an error. Also, change the algorithm used to fold exponentiation to make it
much faster when the exponent is big.

Add a few integer exponent tests.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15289 3a8e244f-b0f2-452b-bcba-4c88e055c3ca