Fixed regression test failure involving premature freeing of a string
One of the regression tests assigns a string to a class member.  The currently
checked in code then frees this string, even though the class instance has not
been destroyed.

The fix is to change the copy function for strings.  Instead of simply returning
its argument, as is done for other primitive types, a copy of the string is
made, so it is no longer a problem to free the original string.  However, this
causes many, many string copies to be created, so these need to be freed
appropriately.  The following changes in insertDestructors.cpp do this:

1) Destructors for both classes and records are modified to free any members
   that are of type string.
2) Calls to the primitive PRIM_STRING_COPY are added to the list of constructor
   calls so the analysis mechanism in insertDestructors.cpp can determine if
   these copies can be safely freed.
3) Additional conditions were added to prevent freeing strings prematurely: if
   a string (actually, any object) is assigned to a function's argument or to 
   an object of type ref to string, it is not freed.

The following regression tests related to memory usage now fail:
   memory/diten/reallocToZero
   memory/figueroa/printFinalMemStat
   memory/shannon/allocAndFree
   memory/shannon/memmaxExceedsLineno2
   memory/shannon/memstatPrint
   memory/shannon/printFinalMemStat
   memory/shannon/reallocZeroSize
On some of these, the maximum amount of memory used throughout the program's
execution is now higher due to the string copies.  In other cases, at the time
the statistics on memory usage is printed, some of the string copies have not
yet been freed.  These regression test failures will be dealt with soon.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15291 3a8e244f-b0f2-452b-bcba-4c88e055c3ca