Tests and a small fix related to slicing with ..
Wojciech turned in what I considered to be a bug (captured in
domain64.chpl) in which using an unbounded range ".." to slice a
64-bit indexed domain resulted in a compiler error due to the fact
that the unbounded range was assumed to have int(32) type and the
domain used int(64) indices.  I got this case working by breaking the
boundsCheck() function into two routines with a where clause so that
the fully unbounded case would be handled separately and trivially
return true (the old code had the same behavior, but since it was part
of a larger function, subsequent assignments in the function resulted
in type mismatch errors).  This got domain64.chpl working.

domain64-workaround.chpl is a workaround to domain64 that I sent
Wojciech before fixing the error.  

domain64-badworkaround.chpl is a future that I'm filing that shows
that an unbounded range can be assigned to a (presumably) bounded
range without any complaints, yet without actually doing the right
thing either.

domain641D.chpl is a 1D version of domain64.chpl.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15292 3a8e244f-b0f2-452b-bcba-4c88e055c3ca