Fixed problem in my previous check in related to freeing strings
The problem was due to an explicit destructor in ChapelIO which was unnecessary
since the compiler creates a default destructor that now frees any members of
type string.  The explicit destructor was causing the string member to be freed
twice.

Changed the line numbers in a .good file (since ChapelIO changed), and
simplyfied the _copy function in ChapelBase.  (There is now just one function
instead of two.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15296 3a8e244f-b0f2-452b-bcba-4c88e055c3ca