Update the _ddata class to allow more than 2**31 - 1 elements
To do this, I removed the size field and instead made it a generic integral
argument of the _ddata.init method. I also changed the argument for _ddata.this
from int to generic integral.

The only uses of the size field were in the init method mentioned above and
in the _ddata.writeThis method. I changed this writeThis method to issue a
compilerError if it gets called.

Updated one test that directly creates a _ddata class.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15303 3a8e244f-b0f2-452b-bcba-4c88e055c3ca