Fixed a customer-reported issue on Sun machines
Apparently the isinf macro is normally defined on Sun machines, but not on ours.
The change is to define this macro only if it is not already defined.

This change has not been tested, since I am not able to build the Chapel
compiler on Solaris - I don't know where to get svn from.  However, it seems
like a safe change.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15310 3a8e244f-b0f2-452b-bcba-4c88e055c3ca