Fixed the bug in the test case Steve checked in today which has to do with having a return statement that is not inside any function. This was fixed by removing the return statement that was added to every init function, and then making the check for return statements outside of functions to apply to any return statement, not just those that actually return a value.
This also has the side benefit of getting rid of a goto in the init functions,
which means that anything that would otherwise be freed were it not for the
presence of this goto can now be freed.

The reason the goto was being inserted is that there was a check at the
beginning of every init function to make sure it was not being executed more
than once.  This check is still needed, but it now appears in all the places
that call init functions, rather than inside the init functions themselves.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15327 3a8e244f-b0f2-452b-bcba-4c88e055c3ca