Updating allocAndFree.good
Updating the output file for allocAndFree to reflect the new
max memory numbers that result from copying and freeing
strings.  This seems worth tracking, but if it ends up being
too fragile across future changes, we can reconsider -- in
particular, I believe we could remove a bunch of writelns() and
make strings empty in order to get back to our previous .good
file.  But let's see how this version goes.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15333 3a8e244f-b0f2-452b-bcba-4c88e055c3ca