Applying recommended GASNet patch
In email conversations with Paul today, he suggested we grab the
following patch as it fixes some seg faults on recent x86_64
architectures.  I'm not sure at this point whether the problems
show up in a spurious fashion (in which case this might explain
some of the nondeterministic seg faults we've been seeing) or
deterministically (in which case it might explain the seg faults
I've been getting on CX1 with the IBV conduit).

The official webpage of the patch is here:

  http://upc.lbl.gov/download/patch/berkeley_upc-2.8.0-patch1.shtml

I added mention of this patch to our README.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15342 3a8e244f-b0f2-452b-bcba-4c88e055c3ca