Trying to remove most svn:ignore options from GASNet directory
Previously, we had set svn:ignore on the GASNet-1.12.0 directory
to include configure and a few other files because Steve kept
checking in changes to them accidentally.  The GASNet team
asserts that this shouldn't be necessary if we build and install
correctly.  I feel like it has been at times, but can't seem
to reproduce it right now, so am disabling these ignores for
the time being to help determine when they show up.

So, if anyone finds that they have modifications in their GASNet
directory as a result of simply building GASNet (or the runtime
or ...?) please let me know what you're doing to get that
behavior.  Thanks.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15345 3a8e244f-b0f2-452b-bcba-4c88e055c3ca