Merging Berkeley cnos/PMI patch into our source tree
Berkeley took David's patch and made it a bit more bulletproof/to their
liking.  In this check-in, I bring their patch into our source tree so
that we're in synch with them in this regard going forward.  I generated
these files by checking out a clean copy of GASNet-1.12.0, applying
the patch in comment #7 from:

http://upc-bugs.lbl.gov/bugzilla/show_bug.cgi?id=2531

and then copying the files back into our GASNet tree.  David, if you
get a chance to review this and make sure it seems appropriate, that
would be great.  I don't claim to understand this code...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15346 3a8e244f-b0f2-452b-bcba-4c88e055c3ca