Adding functional code to PVM launcher.
This commit adds functional code to the PVM launcher modeled after the GASNet launcher. It provides basic functionality, and it allows the user to spawn applications if the Chapel application shares a name with a PVM application. A lot of work needs to be done surrounding number of locales and selecting the correct application. This change also cleans up the tabs.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15350 3a8e244f-b0f2-452b-bcba-4c88e055c3ca