Eliminated race condition related to module initialization
Before this check in, if two different threads tried to call the same module
initialization function, the module could be initialized more than once, since
both threads could read the module guard variable before either one had a
chance to set it to false.  The fix is to have two module guard variables: one
that is replicated on each locale to reduce communication (this was already
there prior to this check in), and a sync variable on locale 0 to ensure only
one thread can read the guard at a time.

This is a bit tricky to implement, because sync variables cannot be used before
a "use ChapelBase" is processed, since ChapelBase defines the operations on
sync variables.  To get around this, a "use ChapelBase" is inserted at the very
top of theProgram's init function.  Next, all the sync variables guarding the
various module init functions are defined.  Then, a "use ChapelStandard" is
inserted.  (There are no longer any guards whatsoever for ChapelStandard's init
function, since only locale 0 can encounter a "use ChapelStandard".  There are
per-locale guards for ChapelBase's init function, but no sync guard, since only
locale 0 can encounter a "use ChapelBase"; ChapelStandard has a (redundant)
"use ChapelBase", so having a per-locale guard prevents calling ChapelBase's
init function twice.)

Whenever any other use statement is processed, it is converted into a check
of the corresponding per-locale guard, which, if successful, causes the guard
to be set to false.  Next, an on statement is inserted to read the sync guard
on locale 0.  If the guard is true, the init function is called; otherwise, the
guard is set to false so it can be read again.  The very first statement in any
init function, other than ChapelStandard's and ChapelBase's, sets the sync
guard to false.  This is done before calling any other module's init function,
for example.

A "hack" was used to insert sync guard declarations before the "use
ChapelStandard", and to make sure calls to init functions inside of another init
function occurred after setting the sync guard to zero.  Perhaps this should be
cleaned up at some later time.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15352 3a8e244f-b0f2-452b-bcba-4c88e055c3ca