Next step in making GASNet build cleanly
I'm continuing to work with the GASNet team to determine how to make
GASNet build cleanly so that it doesn't modify files in the
GASNet-1.12.0 directory.  They suggested the following steps to
ensure that timestamps on the autoconf files are in the correct
order:

touch aclocal.m4
sleep 2
touch configure gasnet_config.h.in
sleep 2
find . -name Makefile.in -print | xargs touch

Apparently gcc uses a similar script for similar purposes.  This
change to the Makefile incorporates this step.  It worked well
on Linux -- next step is to test on other platforms.

This checkin also contains a small change to make our use of
export more-sh- and less-bash-oriented.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15354 3a8e244f-b0f2-452b-bcba-4c88e055c3ca