Fix to our MPI GASNet builds on XT
Steve reported the other week that when he tried to set
CHPL_COMM_SUBSTRATE to mpi he couldn't build GASNet because
the segment is inferred to be "everything" and this breaks
the build of the portals conduit.  After some discussion
with the GASNet team, determined that everything is an
appropriate choice of segment for MPI on the XT and that
portals shouldn't be built when segment everything has been
requested.  This is a patch to the configure.in file
provided by the GASNet team, obtained from:

http://upc-bugs.lbl.gov/bugzilla/show_bug.cgi?id=2544

along with a new configure file generated by re-running
Bootstrap.

Updated our GASNet README to mention this patch and
refactored the text of the list of patches since it
had gotten a bit repetitive.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15359 3a8e244f-b0f2-452b-bcba-4c88e055c3ca