Updated the memmaxExceeds tests again
The change in the limit from 25 to 49 turned out to be too low when the
developer flag is not set, since the line number and filename are also passed
in as part of the arguments for an on clause.  (These are not passed in when
the developer flag is set.)  The limit for both of these tests is now 160,
which is 95 bytes more than needed.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15363 3a8e244f-b0f2-452b-bcba-4c88e055c3ca