Fixed the regressions with --no-copy-propagation
These were due to strings being freed prematurely.  This happened because the
only use was inside a get ref primitive, which prevented my analysis from
seeing the enclosing primitive move.  The fix was to remove a "#if 0" that
enabled seeing the enclosing primitive move.  In addition, the code that
disabled freeing memory when assigning to a variable whose type is the ref type
of the variable to be freed was modified to do this only if variable to be
freed is not in call (such as a get ref primitive).  Just to be safe, these
changes take effect only done for strings, though it would probably work for
other types as well.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15364 3a8e244f-b0f2-452b-bcba-4c88e055c3ca