Removing 64-bit flags from IBM tools
It turns out that because different IBMs use different
versions of ar, the flags I used yesterday are not as
portable as I'd hoped.  In particular, some locations
use gnu ar and others use IBM's.  Instead, an environment
variable can and should be used to specify the 64-bit
build required by GASNet more portably, so I am going
to document that in an IBM-specific README instead.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15369 3a8e244f-b0f2-452b-bcba-4c88e055c3ca