Introduced an array of function pointers in the compiler-generated code called chpl_ftable.  Unified function pointer types in the runtime to the single type chpl_fn_p and introduced chpl_fn_int_p for an integral type that can be used to index into chpl_ftable.  I set this to int16_t for now.
Changed the runtime interface functions that take function pointers to
take chpl_fn_int_t integers except for chpl_begin (which is not called
directly from the generated code).  This removes the assumption that
function pointers are the same on different machines.

Updated all comm and threads layers.  I expect that there are errors
in the ARMCI runtime now that need to be looked into.  I also
commented a concern in the MPI runtime.

Made greprt more robust to additions and changes to the runtime.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15370 3a8e244f-b0f2-452b-bcba-4c88e055c3ca