Updating PVM launcher with recent changes.
This change continues work on the PVM runtime. It still employs the cheat whereby a Chapel program sharing names with a PVM app will simply run the PVM program, though leaving this in place allows the programmer to know if the issued command line works. The way this works is it compiles the list of nodes from the hostfile taking advantage of the fact that adding/deleting the server node issues a non-fatal error in PVM. The command generated is: echo "add <node><newline>...spawn -> <app><newline>delete <node><newline>..." | <path>pvm. Another thing that needs to be cleaned up here is the memory allocation. Right now, it's pretty crude s.t. I could get something running.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15373 3a8e244f-b0f2-452b-bcba-4c88e055c3ca