Fixed memory leak for strings used in conditional statements
Before, if a string is used just once inside of a conditional statement, it
would be freed just after this use, and if memory for the string was allocated
outside of the conditional statement, and the string was never actually used,
this memory would be leaked.  With this check in, if the only use is in a
conditional statement, the string is freed just after the outermost enclosing
conditional statement.

A test case is also included in this check in, though it fails for gasnet and
--no-local.  This test case includes a nested conditional statement, as well as
a select statement.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15376 3a8e244f-b0f2-452b-bcba-4c88e055c3ca