Update to IBM strings and settings
It turns out that power5 and power6 machines have
different networks and preferred GASNet conduits,
so I distinguished between these platforms in our
util scripts using the strings 'pwr5' and 'pwr6'
which are also the strings that IBM uses according
to a colleague.  I still need to update the
documentation but prefer to do that on a local
machine where backspace works properly in emacs. :)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15383 3a8e244f-b0f2-452b-bcba-4c88e055c3ca